Refactor AbstractLicensee
