+++ joylive-core/joylive-governance-api/src/main/java/com/jd/live/agent/governance/invoke/ratelimit/AbstractRateLimiter.java
@@ -19,10 +19,8 @@ import com.jd.live.agent.core.util.option.MapOption;
 import com.jd.live.agent.core.util.option.Option;
 import com.jd.live.agent.governance.invoke.permission.AbstractLicensee;
 import com.jd.live.agent.governance.policy.service.limit.RateLimitPolicy;
-import lombok.Getter;
 
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicBoolean;
 
 /**
  * AbstractRateLimiter is an abstract implementation of the RateLimiter interface,
@@ -50,11 +48,6 @@ public abstract class AbstractRateLimiter extends AbstractLicensee<RateLimitPoli
      * The option that contains additional settings that may affect the behavior of the rate limiter.
      */
     protected final Option option;
-
-    @Getter
-    protected long lastAccessTime;
-
-    protected final AtomicBoolean started = new AtomicBoolean(true);
 
     /**
      * Constructs an instance of the AbstractRateLimiter class with the given rate limit policy and time unit.
@@ -95,13 +88,6 @@ public abstract class AbstractRateLimiter extends AbstractLicensee<RateLimitPoli
         return permits <= 0 || doAcquire(permits, timeout, timeUnit);
     }
 
-    @Override
-    public void close() {
-        if (started.compareAndSet(true, false)) {
-            doClose();
-        }
-    }
-
     /**
      * Try to get some permits within a duration and return the result
      *
@@ -111,13 +97,6 @@ public abstract class AbstractRateLimiter extends AbstractLicensee<RateLimitPoli
      * @return result
      */
     protected abstract boolean doAcquire(int permits, long timeout, TimeUnit timeUnit);
-
-    /**
-     * Closes the limiter.
-     */
-    protected void doClose() {
-
-    }
 
 }
